digital: Convert "sps" parameter from int to float in "correlate_and_sync_cc"
Even though the GRC XML file for correlate_and_sync_cc claimed sps
(samples per symbol) was a float, it was implemented in the block as an
unsigned int.  This commit fixes that, changing sps to a float.

+++ gr-digital/lib/correlate_and_sync_cc_impl.h
@@ -35,7 +35,7 @@ namespace gr {
     {
     private:
       std::vector<gr_complex> d_symbols;
-      unsigned int d_sps;
+      float d_sps;
       float d_thresh;
       kernel::fft_filter_ccc  *d_filter;
 
@@ -43,7 +43,7 @@ namespace gr {
 
     public:
       correlate_and_sync_cc_impl(const std::vector<gr_complex> &symbols,
-                                 unsigned int sps, float threshold=0.9);
+                                 float sps, float threshold=0.9);
       ~correlate_and_sync_cc_impl();
 
       std::vector<gr_complex> symbols() const;
